Move UserInvoiceSetupApiService to service folder so it follows the Stacked architecture.
